pyproject.toml
[build-system]
requires = ["setuptools>=77", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "marine-weather2"
readme = "README.md"
authors = []
license = "AGPL-3.0-or-later"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Django",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
]
requires-python = ">=3.11"
dependencies = [
    "arches>=8.1.0,<8.2.0",
    "arches-querysets[drf] @ git+https://github.com/archesproject/arches-querysets@dev/1.1.x",
    "arches-component-lab @ git+https://github.com/archesproject/arches-component-lab@main",
]
version = "0.0.1"

[dependency-groups]
dev = [
    "black==24.4.2",
    "coverage",
    "django-silk==5.1.0",
    "livereload",
    "pre-commit==3.8.0",
    "sst",
]

[tool.setuptools]
packages = ["marine_weather"]
